+++ apps/token/src/routes/staking/node/your-stake.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { KeyValueTable, KeyValueTableRow } from '@vegaprotocol/ui-toolkit';
-import { formatNumber } from '../../lib/format-number';
-import type { BigNumber } from '../../lib/bignumber';
+import { formatNumber } from '../../../lib/format-number';
+import type { BigNumber } from '../../../lib/bignumber';
 
 export interface YourStakeProps {
   stakeThisEpoch: BigNumber;
